requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
langchain
langchain-community
langchain-huggingface
chromadb
pypdf
pydantic
pydantic-settings
python-dotenv
langchain-text-splitters
sentence-transformers
torch
fastapi
uvicorn
python-multipart
